+++ R/betwfe_class.R
@@ -184,3 +184,87 @@ print.summary.betwfe <- function(x, ...) {
 
 	invisible(x)
 }
+
+#-------------------------------------------------------------------------------
+# Constructor validator (#85). See R/fetwfe_class.R for the design rationale.
+# BETWFE is the hybrid: has `att_selected` and lambda.* slots like fetwfe,
+# but X_ints/y/X_final/y_final/calc_ses are at TOP level like etwfe.
+#-------------------------------------------------------------------------------
+
+.EXPECTED_SLOTS_BETWFE <- c(
+	"att_hat",
+	"att_se",
+	"att_p_value",
+	"att_selected",
+	"catt_hats",
+	"catt_ses",
+	"cohort_probs",
+	"cohort_probs_overall",
+	"catt_df",
+	"beta_hat",
+	"treat_inds",
+	"treat_int_inds",
+	"sig_eps_sq",
+	"sig_eps_c_sq",
+	"lambda.max",
+	"lambda.max_model_size",
+	"lambda.min",
+	"lambda.min_model_size",
+	"lambda_star",
+	"lambda_star_model_size",
+	"X_ints",
+	"y",
+	"X_final",
+	"y_final",
+	"N",
+	"T",
+	"R",
+	"d",
+	"p",
+	"alpha",
+	"calc_ses",
+	"indep_counts_used",
+	"se_type",
+	"y_mean",
+	"response_col_name",
+	"time_var",
+	"unit_var",
+	"treatment",
+	"covs"
+)
+
+#' @title Validate a `betwfe`-classed object's contracts
+#' @keywords internal
+#' @noRd
+.validate_betwfe <- function(x) {
+	cls <- "betwfe"
+	.stop_if_missing_slots(x, .EXPECTED_SLOTS_BETWFE, cls)
+	.check_type_sanity(x, cls, has_alpha = TRUE, has_att_selected = TRUE)
+	.check_se_consistency(x, calc_ses_path = "calc_ses", cls)
+	.check_selection_consistency(x, cls)
+	.check_p_value_na(x, cls)
+	.check_catt_df_shape(x, cls)
+	.check_cohort_probs(x, cls)
+	.check_lambda_monotonicity(x, cls)
+	.assert_contract(
+		length(x$beta_hat) == x$p,
+		"C6 length(beta_hat) == p",
+		cls
+	)
+	.assert_contract(
+		length(x$y) == x$N * x$T,
+		"C6 length(y) == N * T",
+		cls
+	)
+	.assert_contract(
+		nrow(x$X_ints) == x$N * x$T,
+		"C6 nrow(X_ints) == N * T",
+		cls
+	)
+	.assert_contract(
+		is.logical(x$calc_ses) && length(x$calc_ses) == 1L,
+		"C8 calc_ses is length-1 logical",
+		cls
+	)
+	invisible(x)
+}